+++ sdk/internal/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 1.9.0 (Unreleased)
 
 ### Features Added
-* Added `test/credential.New()`, a credential factory for tests
+* Added `test/credential` package containing a credential factory for live and recorded tests
+  and a fake credential for offline tests.
 
 ### Breaking Changes